==> 2_Codebook.ipynb <==
# Executing definitions file to get fundus image details and all function definitions

%run ./1_Definitions.ipynb

## Code Book Creation


**disease_types** ['eye','neuro','cardio','systemic'] as keys and disease codes belonging to that category as values

**surgery_types** same keys as above with surgery/disease codes belonging to that category as values

**disease_codebook** disease code as keys and description as values

# Data Coding 6
k = ['eye','neuro','cardio','systemic']
v = [[1076,1265,1268,1300,1301,1302,1303,1304,1305,1307,1308,1490,1590,1591,1592,1593,1681],
    [1076,1266,1270,1271,1272,1273,1274,1275,1276,1277,1278,1280,1281,1282,1283,1284,1285,1286,1287,1288,1289,1290,1291,1292,1293,1438,1442,1488,1489,1491,1492,1528,1586,1587,1588,1589,1733,1779],
    [1065,1066,1067,1071,1081,1082,1083,1084,1085,1089,1090,1091,1092,1093,1094,1095,1096,1097,1098,1099,1100,1103,1104,1105,1111,1112,1480,1481,1532,1536,1542,1546,1547,1548,1549,1550,1551,1552,1553,1554,1555,1556,1557,1558,1559,1649,1650,1651,1652,1653,1654,1655,1656,1657,1658,1659],
    [1075,1080,1245,1246,1247,1248,1414,1415,1416,1417,1419,1420,1421,1422,1423,1424,1425,1426,1427,1428,1429,1430,1543,1544,1545,1584,1625,1626,1627,1628,1642,1643]]
disease_types = {k[i]: v[i] for i in range(len(k))}

# Data Coding 5
v1 = [[1076,1481,1482,1483,1484,1548,1651,1652],
     [1076,1521,1522,1523,1524,1525,1526,1527,1528],
     [1071,1086,1087,1088,1113,1114,1115,1116,1117,1118,1119,1120,1121,1122,1123,1124,1125,1127,1128,1129,1437,1488,1489,1490,1491,1492,1531,1535,1581,1582,1608,1609,1610,1611,1613,1614,1615],
     []]
surgery_types = {k[i]: v1[i] for i in range(len(k))}

df = pd.read_excel(DATA_dir + 'disease_codebook.xlsx')
disease_codebook = dict([(k,v) for k,v in zip(df.CODE,df.DISEASE)])
disease_codebook[0] = 'nan'

##### Latest Disease codebook by ICD 10

# Using Data Coding 19 from UK Bio Bank Dataset

# Creating disease codebook for latest disease list by ICD10
df = pd.read_excel(DATA_dir + 'ICD10_DC19.xlsx')

# This is a dictionary having the disease name/ meaning for each code defined. key: code, value: meaning
icd10 = dict([(key,val) for key,val in zip(df.coding,df.meaning)])
discode_list = df["coding"].tolist()


# EYE
sub_eye = []
for i in range(7):
    sub_eye.append('H0'+str(i))
for i in range(10,14):
    sub_eye.append('H'+str(i))   
for i in range(15,23):
    sub_eye.append('H'+str(i))
for i in range(25,29):
    sub_eye.append('H'+str(i))
for i in range(30,37):
    sub_eye.append('H'+str(i))    
for i in range(40,43):
    sub_eye.append('H'+str(i))    
for i in range(43,46):
    sub_eye.append('H'+str(i))    
for i in range(46,49):
    sub_eye.append('H'+str(i))    
for i in range(49,53):
    sub_eye.append('H'+str(i))    
for i in range(53,55):
    sub_eye.append('H'+str(i))    
for i in range(55,60):
    sub_eye.append('H'+str(i))


# NEURO
sub_neuro = []
for i in range(10):
    sub_neuro.append('G0'+str(i))
for i in range(10,15):
    sub_neuro.append('G'+str(i))   
for i in range(20,27):
    sub_neuro.append('G'+str(i))
for i in range(30,33):
    sub_neuro.append('G'+str(i))
for i in range(35,38):
    sub_neuro.append('G'+str(i))    
for i in range(40,48):
    sub_neuro.append('G'+str(i))    
for i in range(50,60):
    sub_neuro.append('G'+str(i))    
for i in range(60,65):
    sub_neuro.append('G'+str(i))    
for i in range(70,74):
    sub_neuro.append('G'+str(i))    
for i in range(80,84):
    sub_neuro.append('G'+str(i))    
for i in range(90,100):
    sub_neuro.append('G'+str(i))
    

# CARDIO
sub_cardio = []
for i in range(3):
    sub_cardio.append('I0'+str(i))
for i in range(5,10):
    sub_cardio.append('I'+str(i))   
for i in range(10,16):
    sub_cardio.append('I'+str(i))
for i in range(20,26):
    sub_cardio.append('I'+str(i))
for i in range(26,29):
    sub_cardio.append('I'+str(i))    
for i in range(30,53):
    sub_cardio.append('I'+str(i))    
for i in range(60,70):
    sub_cardio.append('I'+str(i))    
for i in range(70,80):
    sub_cardio.append('I'+str(i))    
for i in range(80,90):
    sub_cardio.append('I'+str(i))    
for i in range(95,100):
    sub_cardio.append('I'+str(i))

sub = [] # list of 3 lists -eye, neuro, cardio
sub.append(sub_eye)
sub.append(sub_neuro)
sub.append(sub_cardio)

mykey = ['EYE','NEURO','CARDIO']
myval = [[],[],[]]

for sub_type in range(3):
    for i in discode_list:
        for j in sub[sub_type]:
            if(str(j) in str(i)):
                myval[sub_type].append(str(i))


# dc19_categories is a dictionary having keys as ['EYE','NEURO','CARDIO'] and their respective diseases as the values in a list.
dc19_categories = {mykey[i]: myval[i] for i in range(len(mykey))}

# feat_desc = pd.read_excel(DATA_dir + 'FeatureNames.xlsx')
# feat_desc_dict = dict([(k,v) for k,v in zip(feat_desc.UDI,feat_desc.Description)])

### Data-coding 19

# Getting the specific columns which use data-coding 19
# We do not consider the type of cancer and external cause of death

df_prim_death = fundus_df.iloc[:,5720:5722] # Underlying (primary) cause of death: ICD10
df_sec_death = fundus_df.iloc[:,5722:5744] # 	Contributory (secondary) causes of death: ICD10
df_main = fundus_df.iloc[:,5885:5951] # Diagnoses - main ICD10
df_sec = fundus_df.iloc[:,5951:6135] # Diagnoses - secondary ICD10
df_19 = pd.concat([df_prim_death, df_sec_death, df_main, df_sec], axis=1)
df19_arr = df_19.to_numpy()

# Removing None elements to form a cleaner list
df19_clean = []
for i in range(df19_arr.shape[0]):
    temp = df19_arr[i]
    df19_clean.append(temp[temp != np.array(None)])    

# Unique element for each row/ patient
df19_clean_unique = []

for i in range(len(df19_clean)):
    temp_list = []
    for j in range(df19_clean[i].shape[0]):
        if (len(temp_list) == 0):
            temp_list.append(df19_clean[i][j])
        elif(df19_clean[i][j] not in temp_list):
            temp_list.append(df19_clean[i][j])
        else:
            continue
    df19_clean_unique.append(temp_list)
    
# Converting the final unique clean list of arrays to 2D numpy array for easier access
df19_clean_arr = np.array([np.array(xi) for xi in df19_clean_unique])

# Count of each unique disease in the dataset
uniqueValues, occurCount = np.unique(np.asarray([item for sublist in df19_clean_arr for item in sublist]), return_counts=True)
# Creating a dictionary to access the frequency of each unique disease
# disease_freq = {uniqueValues[i]: occurCount[i] for i in range(len(occurCount))}
disease_freq = dict(zip(uniqueValues, occurCount))
# Getting the list in a sorted array for viewing purpose
sort_disease_freq = sorted(disease_freq.items(), key=lambda x: x[1], reverse=True)

# Diseases by categories

# For All diseases
discode_cat = [] # has all the major disease codes of data-coding 19
for i in range(len(discode_list)):
    if(len(discode_list[i]) == 3):
        discode_cat.append(discode_list[i])
        
# For Eye, Neuro and Cardio diseases only
discode_list_specific = dc19_categories['EYE'] + dc19_categories['NEURO'] + dc19_categories['CARDIO']
discode_cat_specific = []
for i in range(len(discode_list_specific)):
    if(len(discode_list_specific[i]) == 3):
        discode_cat_specific.append(discode_list_specific[i])
        
count_cat_specific = []
for i in range(len(discode_cat_specific)):
    tempsum = 0
    for j in range(df19_clean_arr.shape[0]):
        if( len(df19_clean_arr[j])> 1 ):
#             print(df_clean_arr[j])
            tempsum = tempsum + np.sum(np.char.count(df19_clean_arr[j], sub = discode_cat_specific[i]))
#             print(tempsum)
    count_cat_specific.append(tempsum)

# Dictionary to store discode_cat_specific and its count count_cat_specific in a asorted manner

dict_specific = {discode_cat_specific[i]: count_cat_specific[i] for i in range(len(count_cat_specific))}
sort_dict_specific = sorted(dict_specific.items(), key=lambda x: x[1], reverse=True)

### Data-coding 100523

df_100523 = fundus_df.iloc[:,4248:4263]
df100523_arr = df_100523.to_numpy()

df100523_unique = []
for i in range(len(df100523_arr)):
    temp_list = []
    for j in range(df100523_arr[i].shape[0]):
#         print(df100523_arr[i][j], temp_list, "\n\n")
        if(~np.isnan(df100523_arr[i][j])):
            if (len(temp_list) == 0):
                temp_list.append(df100523_arr[i][j])
            elif(df100523_arr[i][j] not in temp_list):
                temp_list.append(df100523_arr[i][j])
            else:
                continue
    df100523_unique.append(temp_list)

# Converting the final unique clean list of arrays to 2D numpy array for easier access
df100523_unique_arr = np.array([np.array(xi) for xi in df100523_unique])

disease_freqq = get_sorted_unique(df_100523)
dict_100523 = {1:'Diabetes related eye disease',2:'Glaucoma',3:'Injury or trauma resulting in loss of vision',
               4:'Cataract',5:'Macular degeneration',6:'Other serious eye condition',-7:'None of the above',-3:'Prefer not to answer',-1:'Do not know',0:'NaN'}


# Age at which glaucoma is diagnosed

#Checking count of how many people answered with an age value
df_100291 = fundus_df.iloc[:,1516:1519]
df_100291_arr = df_100291.to_numpy()
df_100291_list = []
for i in range(df_100291_arr.shape[0]):
    temp = df_100291_arr[i]
    if(temp[~np.isnan(temp)].any()):
        df_100291_list.append(temp[~np.isnan(temp)])

# Dictionary of which eye
which_eye_dict = {0: "Don't know", 1: "Left Eye", 2: "Right Eye", 3: "Both Eyes"}

# Which eye affected by Glaucoma
df_100515 = fundus_df.iloc[:,4209:4212]
a = df_100515['6119-0.0'].value_counts()
b = df_100515['6119-1.0'].value_counts()
c = df_100515['6119-2.0'].value_counts()
d = c.add(a.add(b))


# Ever had glaucoma surgery or laser treatment for glaucoma or high eye pressure - data coding 100272
df_100272 = fundus_df.iloc[:,2720:2724]
df_100272_arr = df_100272.to_numpy()
df_100272_list = [] # Answers given by the patient about surgery
df_100272_list_pid = [] # Which patient answered it with 0,2,3,4,5

surgery_patients = [] # patients who replied either with left right or both eye had surgery
surgery_id = []       # which eye 2: right, 3: left, 4: both

dict_100272 = {0: "No Surgery", 2: "Right Eye Operated", 3: "Left Eye Operated", 4: "Both Eyes Operated", 5: "Don't Know"}

for i in range(df_100272_arr.shape[0]):
    temp = df_100272_arr[i]
    if(temp[~np.isnan(temp)].any()):
        df_100272_list.append(temp[~np.isnan(temp)])
        df_100272_list_pid.append(i)


# Patients who said Dont know for 100523 but have glaucoma as per icd10
df_100523_arr = df_100523.to_numpy()
dont_know_patients = [] # Patients who said don't know in questionnaire 100523 - if doctor said they have eye disease

for i in range(len(df_100523_arr)):
    if(-1 in df_100523_arr[i]):
        dont_know_patients.append(i)        

# For These "dont_know_patients", we will check in icd10, if they were diagnosed glaucoma

dont_know_glauicd10 = [] # Patients who said don't know in questionnaire but found to have glaucoma as per diagnosis in icd10

for pat in range(len(dont_know_patients)):
    test_list = df_19.iloc[dont_know_patients[pat]].unique()
    for i in test_list:
        if not i:
            continue
        elif pd.isnull(i):
            continue
        elif ('H40' in i or 'H42' in i):
#             print(test_list)
            dont_know_glauicd10.append(pat)
            break

print("Codebook Executed !!\n")
