==> 3_Frequent diseases.ipynb <==
%run ./2_Codebook.ipynb

# Data Coding 6 (Non cancer illness code)

# Extracting only those fields that uses Data-coding 6
mydf = fundus_df.iloc[:,4670:4772]
temp_dis = get_sorted_unique(mydf)
disease_specific = disease_types['eye'] + disease_types['neuro'] + disease_types['cardio'] + disease_types['systemic']
len(disease_specific)

##### 1. Counting a disease for each patient only once

# EYE, NEURO, CARDIO and SYSTEMIC diseases only

print("Most frequently occuring diseases are:\n")
for i in range(len(temp_dis)):
    if(temp_dis[i][0] in disease_specific):
        print('{:<50s}{:->10}'.format(disease_codebook[temp_dis[i][0]],temp_dis[i][1]))


##### 2. Counting overall disease occuring frequency

# NOT RELEVANT

df_int = conv2int(mydf)    
data_int = df_int.to_numpy()
x,y = np.unique(data_int, return_counts=True)
x = x[1:]
y = y[1:]
disease_freq_all = {x[i]: y[i] for i in range(len(x))} 
sort_disease_freq_all = sorted(disease_freq_all.items(), key=lambda a: a[1], reverse=True)
print("Most frequently occuring diseases are:\n")
for i in range(50):
    print('{:<50s}{:->10}'.format(disease_codebook[sort_disease_freq_all[i][0]],sort_disease_freq_all[i][1]))


# Data Coding 5  (Non cancer illness surgery)

# Extracting only those fields that uses Data-coding 6
mydf = fundus_df.iloc[:,4772:4868]
temp_dis = get_sorted_unique(mydf)

print("Most frequently performed operations are:\n")
for i in range(len(temp_dis)):
    print('{:<50s}{:->10}'.format(disease_codebook[temp_dis[i][0]],temp_dis[i][1]))


======================
# |Eye Diseases |
======================

# Data-coding 100523 
"Has a doctor told you that you have any of the following problems with your eyes? (You can select more than one answer)"

# Self reported eye diseases

for i in range(len(dict_100523)):
    if ('None' not in dict_100523[disease_freqq[i][0]]) and ('NaN' not in dict_100523[disease_freqq[i][0]]) 
    and ('know' not in dict_100523[disease_freqq[i][0]]):
        print('{:<50s}{:->20}'.format(dict_100523[disease_freqq[i][0]],disease_freqq[i][1]))


### Further questions asked based on above answers: - For Glaucoma

print("1. Age at which glaucoma is diagnosed")

cnt = len(df_100291_list)
print("No. of patients gave an answer:",cnt)
for i in range(len(df_100291_list)):
    for j in range(len(df_100291_list[i])):
        if(df_100291_list[i][j] < 4):
            cnt-=1
print("No. of patients gave an age:",cnt)
# 1653 people have answered with a age for themselves when glaucoma was detected for them.
print("No. of people replied with don't know or prefer not to answer:",len(df_100291_list) - cnt)  


print("2. Which eye(s) affected by glaucoma - data coding 100515")
df_100515 = fundus_df.iloc[:,4209:4212]
a = df_100515['6119-0.0'].value_counts()
b = df_100515['6119-1.0'].value_counts()
c = df_100515['6119-2.0'].value_counts()
d = c.add(a.add(b))
print("Count: 1:Left 2:Right 3:Both")
print(d)
print("Total =", d.sum(), "patients answered with which eye was affected.")

print("3. Ever had glaucoma surgery or laser treatment for glaucoma or high eye pressure - data coding 100272")
      
cnt0 = 0 # Didn't have surgery    
cnt1 = 0 # Don't know
cnt2 = 0 # Had surgery in left, right or both eyes

for i in range(len(df_100272_list)):
    if(0 in df_100272_list[i]):
        cnt0 +=1
    elif(5 in df_100272_list[i]):
        cnt1 +=1
    else:
        cnt2+=1
        surgery_patients.append(df_100272_list_pid[i])
        surgery_id.append(df_100272_list[i])

print("Patients having either left or right or both eyes operated for glaucoma:",cnt2)
print("Patients don't know:",cnt1)
print("Patients who didn't have surgery:",cnt0)

## Which eye had Glaucoma and which eye is operated for a particular patient

for i in range(len(surgery_patients)):
    pat = surgery_patients[i]
    eye_operated = surgery_id[i][-1]
    print("\n\nFor Patient:",pat,"\nGlaucoma found in:",which_eye_dict[which_eye(pat)],"\n", dict_100523[eye_operated])

# Data-coding 100523 


## Displaying most frequent diseases with the count (For each Category) - only includes eye, neuro and cardio diseases 

# Displaying most frequent diseases with the count
for i in range(len(sort_dict_specific)):
    print('{:<50s}{:->20}'.format(icd10[sort_dict_specific[i][0]],sort_dict_specific[i][1]))

## Displaying most frequent diseases with the count (For each individual disease at lowest level)
# Displaying most frequent diseases with the count
for i in range(1,10):
    print('{:<40s}{:->10}'.format(icd10[sort_disease_freq[i][0]],sort_disease_freq[i][1]))


print("Patients who answered Dont know for Glaucoma, but were detected to have glaucoma as per ICD10: ",
      len(dont_know_glauicd10))

=========================================
# |Cardiovascular diseases |
=========================================

# Data-coding 100605

# Vascular/heart problems diagnosed by doctor
df_100605 = fundus_df.iloc[:,4263:4275] #data-coding 100605
disease_freqq = get_sorted_unique(df_100605)
dict_100605 = {1:'Heart attack',2:'Angina',3:'Stroke',4:'High blood pressure',0:'Nan',-3:'Prefer not to answer',-7:'None of the above'}
for i in range(len(dict_100605)):
    print('{:<30s}{:->20}'.format(dict_100605[disease_freqq[i][0]],disease_freqq[i][1]))


# "Doctor restricts physical activity due to heart condition"
df_100267 = fundus_df.iloc[:,4171:4173]
data = df_100267.to_numpy()
np.count_nonzero(data == 1) # count of YES - people those who have heart problems
