==> 4_GroundTruth.ipynb <==
%run ./2_Codebook.ipynb

# Ground Truth

**0:** no glaucoma

**1:** glaucoma

**99:** DON'T KNOW

## Data-Coding 100523

# Using function defitions to get the list of respective ground truths.
GT_100523 = GT_100523()

# Ground truth for left and right eyes separately

GT523_L = [0] * len(fundus_df) # 1525
GT523_R = [0] * len(fundus_df) # 1513

for i in range(len(GT_100523)):
    
    # if patient said he has glaucoma
    if(GT_100523[i] == 1):
        if (which_eye(i) == 1): # left ?
            GT523_L[i] = 1
        elif (which_eye(i) == 2): # right ?
            GT523_R[i] = 1
        elif(which_eye(i) == 3): # both ?
            GT523_L[i] = 1
            GT523_R[i] = 1
        else: # dont know which eye has glaucoma ?
            GT523_L[i] = 1
            GT523_R[i] = 1
    # if patient said don't know or prefer not to answer about glaucoma
    else:
        GT523_L[i] = 1
        GT523_R[i] = 1

## Data-Coding 19 / ICD10 
Since ICD10 does not have classification of which eye had glaucoma, we assumed a patient having glaucoma has it in both eyes.

GT_19 = GT_19()

## Intersection 100523 - ICD10
**0:** no glaucoma in both

**1:** glaucoma in both

**2:** glaucoma in icd10 only

**3:** glaucoma in 100523 only

**99:** DON'T KNOW

GT_intersect = GT_intersect()

GTi_L = [0] * len(fundus_df)
GTi_R = [0] * len(fundus_df)

for i in range(len(GT_intersect)):
    if(GT_intersect[i] == 1): # glaucoma as per 100523 questionnaire, so ask which eye ?
        if(which_eye(i) == 1):
            GTi_L[i] = 1
        elif(which_eye(i) == 2):
            GTi_R[i] = 1
        elif(which_eye(i) == 3):
            GTi_L[i] = 1
            GTi_R[i] = 1
        else:
            GTi_L[i] = 1
            GTi_R[i] = 1


## Union 100523 - ICD10 
**1:** Glaucoma if atleast one of the datasets indicates glaucoma

**0:** None of them indicates glaucoma

**99:** If ICD10 says no glaucoma, and 100523 says doesnt know or prefer not to answer

GT_union = GT_union()

GTu_L = [0] * len(fundus_df)
GTu_R = [0] * len(fundus_df)

for i in range(len(GT_union)):
    if(GT_union[i] == 1 and which_eye(i)!=0):
        if(which_eye(i)==1):
            GTu_L[i] = 1
        elif(which_eye(i)==2):
            GTu_R[i] = 1
        elif(which_eye(i)==3):
            GTu_L[i] = 1
            GTu_R[i] = 1
    elif(GT_union[i] == 1 and which_eye(i)==0):
        GTu_L[i] = 1
        GTu_R[i] = 1
    elif(GT_union[i] == 99):
        GTu_L[i] = 1
        GTu_L[i] = 1


print("Ground truths created are:\n \nGT_100523: GT523_L and GT523_R \nGT_19 \nGT_intersect: GTi_L and GTi_R \nGT_union: GTu_L and GTu_R")

for i in range(len(GT_intersect)):
    if(GT_intersect[i]!=0 and GT_intersect[i]!=1):
        GT_intersect[i]=0

for i in range(len(GT_union)):
    if(GT_union[i]!=0 and GT_union[i]!=1):
        GT_union[i]=1

for i in range(len(GT_100523)):
    if(GT_100523[i]!=0 and GT_100523[i]!=1):
        GT_100523[i]=1

# Saving Different Ground truths

np.save(DATA_dir + 'GroundTruths/GT_intersect.npy', GT_intersect)
np.save(DATA_dir + 'GroundTruths/GT_union.npy', GT_union)
np.save(DATA_dir + 'GroundTruths/GT_19.npy', GT_19)
np.save(DATA_dir + 'GroundTruths/GT_100523.npy', GT_100523)

# Saving Different Ground truths -  for left and right eyes separately

np.save(DATA_dir + 'GroundTruths/GTu_L.npy', GTu_L)
np.save(DATA_dir + 'GroundTruths/GTu_R.npy', GTu_R)

np.save(DATA_dir + 'GroundTruths/GTi_L.npy', GTi_L)
np.save(DATA_dir + 'GroundTruths/GTi_R.npy', GTi_R)

np.save(DATA_dir + 'GroundTruths/GT523_L.npy', GT523_L)
np.save(DATA_dir + 'GroundTruths/GT523_R.npy', GT523_R)
