==> 6_Fundus image extraction.ipynb <==
This file is used only for fundus image extraction for train, validation and test sets respectively 

**cropped**

**resized**

**saved** 

into directory for easier access

%run ./1_Definitions.ipynb

# Whcih GPU?
os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"   # so the IDs match nvidia-smi
os.environ["CUDA_VISIBLE_DEVICES"] = "7"   # choose here your GPU (0-7)

# PATH
srcpath = '/ds2/data/retina/UK_BB/UK_BB_DATA/fundus/'

## Saving Fundus image file names - left and Right

subdirs, dirs, files = os.walk(srcpath).__next__()
m = len(files)

# #SAVED
# filenames_L = []
# filenames_R = []
# for i in range(len(fundus_df)):
#     filenames_L.append(get_img_name(i,df_fundus_left))
#     filenames_R.append(get_img_name(i,df_fundus_right))
    
# np.save(DATA_dir+'filenames_L.npy', filenames_L)
# np.save(DATA_dir+'filenames_R.npy', filenames_R)

## Loading Train Test Validation - Fundus image names and Ground truth

# FUNDUS IMAGE NAMES AND GROUND TRUTHS - LEFT and RIGHT
fundus_train = pd.read_pickle(tts_path + 'fundus_train.pkl')
fundus_val = pd.read_pickle(tts_path + 'fundus_val.pkl')
fundus_test = pd.read_pickle(tts_path + 'fundus_test.pkl')

fundus_train.shape, fundus_val.shape, fundus_test.shape

fundus_train.head()

### Cropping and Resizing images and saving in another folder - LEFT /RIGHT , Train/Test/Validation
# DONE !!

# img_size = 250
# sett = ['val','train','test']
# fcols={0:'left',5:'right'}

# for n,s in enumerate(sett):
#     print(n,": For", s, "Image set.")
#     destpath = DATA_dir + 'Fundus_images_' + str(img_size) + '_CR/' + s + '/'
#     print("Destination path:", destpath,"\n")
#     curr_df = eval('fundus_' + s)
#     L = len(curr_df)
    
#     for k in fcols.keys():
#         print("\t Processing", fcols[k], "fundus images:\n")
        
#         cnt = 0
#         for i in range(L):
#             img_f = curr_df[curr_df.columns[k]].iloc[i]
#             if(os.path.isfile(srcpath + img_f)):
#                 image = imageio.imread(srcpath+ img_f)
#                 if(len(image.shape)!=3):
#                     image = np.repeat(image[:, :, np.newaxis], 3, axis=2)   
#                     print("Image converted from 2D to 3D:",img_f)
#                 ROI = get_ROI(image)
#                 imResize = cv2.resize(ROI, dsize=(img_size, img_size))
#                 matplotlib.image.imsave(destpath + img_f, imResize)
#                 cnt+=1
    
#         print("\t\t",cnt, fcols[k], s, " images resized and stored \n")
